Moved fetchuserdata to settings actions as we no longer need it elsewhere
