+++ contentcuration/contentcuration/frontend/shared/vuex/session/index.js
@@ -1,4 +1,3 @@
-import throttle from 'lodash/throttle';
 import client from '../../client';
 import Languages from 'shared/leUtils/Languages';
 
@@ -20,24 +19,6 @@ function langCode(language) {
     return language.toLowerCase();
   }
 }
-
-const throttleTime = 30 * 1000;
-
-const deferredUser = throttle(
-  function() {
-    return client.get(window.Urls.deferred_user_data());
-  },
-  throttleTime,
-  { trailing: false }
-);
-
-const settingsDeferredUser = throttle(
-  function() {
-    return client.get(window.Urls.deferred_user_data(), { params: { settings: true } });
-  },
-  throttleTime,
-  { trailing: false }
-);
 
 export default {
   state: () => ({
@@ -71,11 +52,14 @@ export default {
     currentUserId(state) {
       return state.currentUser.id;
     },
-    availableSpace(state) {
-      return state.currentUser.available_space || null;
+    usedSpace(state) {
+      return state.currentUser.disk_space_used;
     },
     totalSpace(state) {
       return state.currentUser.disk_space;
+    },
+    availableSpace(state, getters) {
+      return getters.totalSpace - getters.usedSpace;
     },
     storageUseByKind(state) {
       return state.currentUser.space_used_by_kind || null;
@@ -98,25 +82,12 @@ export default {
     updateFullName(context, { first_name, last_name }) {
       let currentUser = context.state.currentUser;
       currentUser = { ...currentUser, first_name, last_name };
-      context.commit('SET_CURRENT_USER', currentUser);
+      context.commit('UPDATE_CURRENT_USER', currentUser);
     },
-    fetchDeferredUserData(context, settings = false) {
-      if (context.getters.availableSpace) {
-        if (
-          (context.getters.storageUseByKind && context.state.currentUser.api_token) ||
-          !settings
-        ) {
-          return;
-        }
-      }
-      let promise;
-      if (settings) {
-        promise = settingsDeferredUser();
-      } else {
-        promise = deferredUser();
-      }
-      return promise.then(response => {
-        context.commit('UPDATE_CURRENT_USER', response.data);
+    fetchUserStorage(context) {
+      return client.get(window.Urls.user_get_storage_used()).then(response => {
+        context.commit('UPDATE_CURRENT_USER', { disk_space_used: response.data });
+        return response.data;
       });
     },
   },
